+++ app/db/core/src/commonMain/kotlin/com/moneymanager/database/repository/CurrencyRepositoryImpl.kt
@@ -62,7 +62,7 @@ class CurrencyRepositoryImpl(
             queries.update(
                 code = currency.code,
                 name = currency.name,
-                scaleFactor = currency.scaleFactor.toLong(),
+                scaleFactor = currency.scaleFactor,
                 id = currency.id.toString(),
             )
         }
